Fix the FDI module
- Fix a numerical issue: numpy ``isclose`` method is used instead of
  direct comparison with ``0``.
- Move the ``main.py`` scipt to ``test`` directory
