+++ test/grouping.py
@@ -8,36 +8,14 @@ import fym.logging
 from ftc.models.multicopter import Multicopter
 from ftc.agents.backstepping import BacksteppingController
 from ftc.agents.grouping import Grouping
+from ftc.agents.fdi import SimpleFDI
 from ftc.faults.actuator import LoE, LiP, Float
-
-
-class FDI(BaseSystem):
-    def __init__(self, no_act, tau):
-        super().__init__(np.eye(no_act))
-        self.tau = tau
-
-    def get_true(self, u, uc):
-        w = np.hstack([
-            ui / uci
-            if (ui != 0 and uci != 0) else 1
-            if (ui == 0 and uci == 0) else 0
-            for ui, uci in zip(u, uc)])
-        return np.diag(w)
-
-    def get_index(self, W):
-        fault_index = np.where(np.diag(W) != 1)[0]
-        return fault_index
-
-    def set_dot(self, W):
-        What = self.state
-        self.dot = - 1 / self.tau * (What - W)
 
 
 class Env(BaseEnv):
     def __init__(self):
         super().__init__(dt=0.01, max_t=20)
         self.plant = Multicopter()
-        # self.controller = BacksteppingController(self.plant.pos.state, self.plant.m, self.plant.g)
 
         # Define faults
         self.sensor_faults = []
@@ -45,11 +23,10 @@ class Env(BaseEnv):
             LoE(time=3, index=0, level=0.5),
             LoE(time=5, index=1, level=0.2),
             LoE(time=7, index=2, level=0.5),
-            # Float(time=10, index=0),
         ]
 
         # Define FDI
-        self.fdi = FDI(no_act=self.plant.mixer.B.shape[1], tau=0.1)
+        self.fdi = SimpleFDI(no_act=self.plant.mixer.B.shape[1], tau=0.1)
         self.grouping = Grouping(self.plant.mixer.B)
 
     def step(self):
@@ -60,12 +37,10 @@ class Env(BaseEnv):
         x = self.plant.state
         What = self.fdi.state
 
-        # u, W, _, Td_dot, xc, *_ = self._get_derivs(t, x, What)
         u, W, *_ = self._get_derivs(t, x, What)
 
         self.plant.set_dot(t, u)
         self.fdi.set_dot(W)
-        # self.controller.set_dot(Td_dot, xc)
 
     def get_forces(self, x):
         return np.vstack((self.plant.m * self.plant.g, 0, 0, 0))
@@ -85,12 +60,6 @@ class Env(BaseEnv):
 
         f = self.get_forces(x)
         u = u_command = self.control_allocation(f, What)
-        # FM, Td_dot = self.controller.command(
-        #     *self.plant.observe_list(), *self.controller.observe_list(),
-        #     self.plant.m, self.plant.J, np.vstack((0, 0, self.plant.g)),
-        # )
-        # pos_c = np.zeros((3, 1))  # TODO: position commander
-        # u = u_command = self.control_allocation(FM, What)
 
         # Set actuator faults
         for act_fault in self.actuator_faults:
@@ -98,7 +67,6 @@ class Env(BaseEnv):
 
         W = self.fdi.get_true(u, u_command)
 
-        # return u, W, u_command, Td_dot, pos_c
         return u, W, u_command
 
     def logger_callback(self, i, t, y, *args):
@@ -107,9 +75,6 @@ class Env(BaseEnv):
         What = states["fdi"]
         u, W, uc = self._get_derivs(t, x, What)
         return dict(t=t, x=x, What=What, u=u, uc=uc, W=W)
-        # x_controller = states["controller"]
-        # u, W, uc, Td_dot, pos_c, *_ = self._get_derivs(t, x, What)
-        # return dict(t=t, x=x, What=What, u=u, uc=uc, W=W, x_controller=x_controller, pos_c=pos_c)
 
 
 def run():
@@ -147,8 +112,10 @@ def exp1_plot():
 
     plt.show()
 
+
 def exp2():
     run()
+
 
 def exp2_plot():
     data = fym.logging.load("data.h5")
